+++ src/gkernel/dtype/nongeometric/matrix/__init__.py
@@ -1,1 +1,2 @@
 from .complex import *
+from .primitive import *
