+++ src/mkernel/__init__.py
@@ -1,0 +1,2 @@
+from .gkernel_wrapper import *
+from .model import Model
