+++ src/mkernel/gkernel_wrapper.py
@@ -7,23 +7,7 @@ class Ray(pg.Ray, Shape):
 
 
 class Pnt(pg.Pnt, Shape):
-
-
-# @classmethod
-# def is_point_like(cls, *points):
-#     """
-#     check if given iterable can be interpreted as point coordinate
-#     :param points: iterable describing coordinate
-#     :return:
-#     """
-#     for p in points:
-#         if isinstance(p, (list, tuple)) & len(p) == 3:
-#             pass
-#         elif isinstance(p, np.ndarray) & p.shape == (3,):
-#             pass
-#         else:
-#             return False
-#     return True
+    pass
 
 
 class Vec(pg.Vec, Shape):
