+++ src/mkernel/model.py
@@ -1,5 +1,5 @@
 from gkernel.dtype.geometric.primitive import Pln
-
+from gkernel.tools.intersector import Intersector as intx
 
 class ModelIterator:
     def __init__(self, shapes):
@@ -22,14 +22,14 @@ class Model:
     def append_shape(self, shape):
         self._shapes.append(shape)
 
-    def iter(self):
+    def iterator(self):
         """
         iter all shapes in model
         :return:
         """
         for shape in self._shapes:
-            if isinstance(shape, Model):    # if its a sub model iter it
-                for child_shape in shape.iter():
+            if isinstance(shape, Model):  # if its a sub model iter it
+                for child_shape in shape.iterator():
                     yield child_shape
             else:
                 yield shape
@@ -40,6 +40,12 @@ class Model:
         :param ray: to intersect with
         :return:
         """
+        for shape in self.iterator():
+            print(intx.intx(shape, ray))
 
+    def render(self):
+        """
+        redner all shaped inside model
 
-
+        :return:
+        """
